Отремонтировал клиента windows

+++ MSChat/Client_chat/Chats_main.cs
@@ -840,25 +840,32 @@ namespace Client_chat
                         {
                             if (cell.Value.ToString().Contains(dataGridViewChat.SelectedCells[0].Value.ToString()))
                             {
-                                cell.Style.BackColor = Color.Blue;
-                                MessageBox.Show("Сообщения друга нельзя редактировать!");
+                               if( cell.Style.ForeColor == Color.Blue)
+                               {
+                                    MessageBox.Show("Сообщения друга нельзя редактировать!");
+                               }
+                               else
+                               {
+                                    //Поиск dataGridViewChat.SelectedCells[0].RowIndex 
+                                    //Выбераем ячейку сообщение
+                                    int selectedrowindexs = dataGridViewChat.SelectedCells[0].RowIndex;
+                                    //Находим по фильтру id dataGridViewChat
+                                    MessСhat tt = allChat[selectedrowindexs];
+                                    //Передаем значение для редактирования
+                                    textBox1.Text = tt.Message;
+                                    //id редактированного сообщения
+                                    Update_id = tt.Id;
+                                    //ДАЕМ РАЗРЕШЕНИЕ ПРИ РЕДАКТИРОВАНИЕИ ОТПРАВКУ
+                                    Update_Message = true;
+                                }                           
                             }
                             else
                             {
 
-                                //Поиск dataGridViewChat.SelectedCells[0].RowIndex 
-                                //Выбераем ячейку сообщение
-                                int selectedrowindexs = dataGridViewChat.SelectedCells[0].RowIndex;
-                                //Находим по фильтру id dataGridViewChat
-                                MessСhat tt = allChat[selectedrowindexs];
-                                //Передаем значение для редактирования
-                                textBox1.Text = tt.Message;
-                                //id редактированного сообщения
-                                Update_id = tt.Id;
-                                //ДАЕМ РАЗРЕШЕНИЕ ПРИ РЕДАКТИРОВАНИЕИ ОТПРАВКУ
-                                Update_Message = true;
+                             
                             }
                             break;
+
                         }
                     }
                 }
@@ -895,47 +902,53 @@ namespace Client_chat
                         {                        
                             //Пишим текущему пользователю что сообщения друга нельзя трогать!
 
-                            cell.Style.BackColor = Color.Blue;
-                            MessageBox.Show("Сообщения друга нельзя удалять!");
+                            if(cell.Style.ForeColor == Color.Blue)
+                            {
+                                MessageBox.Show("Сообщения друга нельзя удалять!");
+                            }
+                            else
+                            {
+                                int selectedrowindexs = dataGridViewChat.SelectedCells[0].RowIndex;
+                                //Находим по фильтру id dataGridViewChat
+                                MessСhat tt = allChat[selectedrowindexs];
+                                //Сообщения для удаления
+                                textBox1.Text = tt.Message;
+                                //id сообщения для удаления
+                                Update_id = tt.Id;
+                                //Для отправки
+                                string FileFS = "";
+                                using (MemoryStream fs = new MemoryStream())
+                                {
+                                    //Определяем текущию дату
+                                    DateTime dateTime = DateTime.Now;
+                                    //Заполняем класс
+                                    MessСhat Mes_chat = new MessСhat(Update_id, Users, Friends, textBox1.Text, dateTime, 1);
+                                    //Серилизуем класс
+                                    JsonSerializer.Serialize<MessСhat>(fs, Mes_chat);
+                                    //Декодируем в строку
+                                    FileFS = Encoding.Default.GetString(fs.ToArray());
+                                }
+                                //Отчищаем
+                                textBox1.Text = "";
+                                //Отправляем на сервер м получаем
+                                using (MemoryStream Delete_dispons = new MemoryStream())
+                                {
+                                    //Отправляем на сервер и там удаляеться сообщение
+                                    Task.Run(async () => await command.Delete_message_make_up(IP_ADRES.Ip_adress, FileFS, "011")).Wait();
+                                    //Отчищаем dataGridViewChat
+                                    dataGridViewChat.Rows.Clear();
+                                    //Заполняем чат 
+                                    Chat(command);
+                                }
+                            }
+                            break;
 
                         }
                         else
                         {
-                            //получает id dataGridViewChat
-                            int selectedrowindexs = dataGridViewChat.SelectedCells[0].RowIndex;
-                            //Находим по фильтру id dataGridViewChat
-                            MessСhat tt = allChat[selectedrowindexs];
-                            //Сообщения для удаления
-                            textBox1.Text = tt.Message;
-                            //id сообщения для удаления
-                            Update_id = tt.Id;
-                            //Для отправки
-                            string FileFS = "";
-                            using (MemoryStream fs = new MemoryStream())
-                            {
-                                //Определяем текущию дату
-                                DateTime dateTime = DateTime.Now;
-                                //Заполняем класс
-                                MessСhat Mes_chat = new MessСhat(Update_id, Users, Friends, textBox1.Text, dateTime, 1);
-                                //Серилизуем класс
-                                JsonSerializer.Serialize<MessСhat>(fs, Mes_chat);
-                                //Декодируем в строку
-                                FileFS = Encoding.Default.GetString(fs.ToArray());
-                            }
-                            //Отчищаем
-                            textBox1.Text = "";
-                            //Отправляем на сервер м получаем
-                            using (MemoryStream Delete_dispons = new MemoryStream())
-                            {
-                                //Отправляем на сервер и там удаляеться сообщение
-                                Task.Run(async () => await command.Delete_message_make_up(IP_ADRES.Ip_adress, FileFS, "011")).Wait();
-                                //Отчищаем dataGridViewChat
-                                dataGridViewChat.Rows.Clear();
-                                //Заполняем чат 
-                                Chat(command);
-                            }
+                            //получает id dataGridViewChat                        
                         }
-                        break;
+                     
                     }
                 }
             }
